abstract common instance vars in carstate init and a generic gear parser static method
