+++ selfdrive/car/subaru/carcontroller.py
@@ -53,7 +53,7 @@ class CarController():
       apply_steer = apply_std_steer_torque_limits(new_steer, self.apply_steer_last, CS.steer_torque_driver, P)
       self.steer_rate_limited = new_steer != apply_steer
 
-      lkas_enabled = enabled and not CS.steer_not_allowed
+      lkas_enabled = enabled
 
       if not lkas_enabled:
         apply_steer = 0
